vold: Fix HW FDE OTA support on SW FDE encrypted device
If an OTA is delivered to a device encrypted with SW FDE, an OTA
of HW FDE should not force user to switch to HW FDE and hence wipe
of all userdata. Instead, it should continue using SW FDE until user
itself perform factory reset. After factory reset, HW FDE would be
used to encrypt the data. This support somehow was broken. This
change fixes the support.

Change-Id: If31ccbd377e5a161a90b2e4a2c53dcc52712b49c